+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # SoftMocks v1 Change Log
 
+## master
+
+There are next changes:
+
+- added tests for constants redefine
+
 ## v1.2.0
 
 There are next changes:
+
 - added Travis and Scrutinizer support
 - skipped running PHP7.0 tests on previously versions of PHP
 - changed default namespace to \Badoo. \QA namespace marked as deprecated and will be removed in 2.0.0